Use the new architecture for the program
Instead of using the Rich library, I started using the Textual framework.
Because the Rich library did not allow you to enter an answer in the input field and change the interface at the same time, and Textual has this capability.

The “backbone” of the game process interface has been created.
The SidebarWidget and GameScreenWidget classes are empty for now, but they are present to create a layout.
Also added a few frames around the containers to make them better visible and easier to work with.