Use the screens
All code for the game process layout is now located in the game_process.py file in the screens folder.
Because using screens allows you to make different “states” of the program, between which you can easily switch.

Also, now all widgets will be located in folders with the same name as the file name for the screens.
In this case, all files with widgets have been moved to the game_process folder and are used in the game_process.py file (screens folder).
This will give an additional structure to the widgets depending on their belonging to a particular screen.

And in general, it will be easier to develop :)