Separate emulated terminal
I created the file “terminal.py”, which now contains the base class for the emulated terminal.
Each subsequent terminal will inherit the base class.

This will make the code less repetitive when the terminal class is used on screens.

P.S. These terminals are necessary for the window to have a given size and not be distorted.