Add 2 widgets to the game creation screen
The first widget I added was a widget for creating a game name.
This will be useful later when connecting to someone's game by name.

The second widget is used to select the amount of the entrance fee.
This will allow users to choose how “valuable” their game will be.

I also added “+” and “-” buttons that increase/decrease the entrance fee by 100. But the minimum is 100.
This will allow you to easily and quickly change the entrance fee, for example, by 200, 300 or 500.