Add a screen for selecting the number of players
A completely new screen that allows you to search
for players (in the future) and choose how many
players will participate in the game.

The functionality of the “CancelButton” and
“CreateButton” buttons has been redesigned. Now
they are "BackButton" and "NextButton" buttons that
can be used repeatedly.

Also some other changes so that the code is not repeated.